fix: member id 변경 시 발생하는 에러 수정 및 주석 처리

+++ bible/src/main/java/com/library/bible/rent/repository/IRentHistoryRepository.java
@@ -24,13 +24,13 @@ public interface IRentHistoryRepository {
 	
 	// RentHistoryRespone
 	List<RentHistoryResponse>selectRentHistoryResponses(
-			@Param("memId") int memId, 
+			@Param("memId") long memId, 
 	        @Param("rentStatus") String rentStatus,
             @Param("offset") int offset, 
             @Param("limit") int limit
     );
 	int countRentHistory(			
-			@Param("memId") int memId, 
+			@Param("memId") long memId, 
 	        @Param("rentStatus") String rentStatus
     );
 }
